chore: refactor names, simplify code
