+++ packages/x-adapter-platform/src/endpoint-adapters/query-suggestions.endpoint-adapter.ts
@@ -1,8 +1,9 @@
 import { endpointAdapterFactory } from '@empathyco/x-adapter-next';
 import { QuerySuggestionsRequest, QuerySuggestionsResponse } from '@empathyco/x-types';
-import { querySuggestionsRequestMapper } from '../mappers/request/query-suggestions-request.mapper';
 // eslint-disable-next-line max-len
-import { querySuggestionsResponseMapper } from '../mappers/response/query-suggestions-response.mapper';
+import { querySuggestionsRequestMapper } from '../mappers/requests/query-suggestions-request.mapper';
+// eslint-disable-next-line max-len
+import { querySuggestionsResponseMapper } from '../mappers/responses/query-suggestions-response.mapper';
 
 export const querySuggestionsEndpointAdapter = endpointAdapterFactory<
   QuerySuggestionsRequest,
