+++ packages/x-adapter-platform/src/endpoint-adapters/search.endpoint-adapter.ts
@@ -1,8 +1,8 @@
 import { endpointAdapterFactory, EndpointAdapterOptions } from '@empathyco/x-adapter-next';
 import { SearchResponse } from '../types/response.types';
 import { SearchRequest } from '../types/request.types';
-import { searchResponseMapper } from '../mappers/response/search-response.mapper';
-import { searchRequestMapper } from '../mappers/request/search-request.mapper';
+import { searchResponseMapper } from '../mappers/responses/search-response.mapper';
+import { searchRequestMapper } from '../mappers/requests/search-request.mapper';
 
 const adapterOptions: EndpointAdapterOptions<SearchRequest, SearchResponse> = {
   endpoint: 'https://api.{env(.)}empathy.co/search/v1/query/{instance}/search',
