+++ packages/x-adapter-platform/src/mappers/__tests__/response/next-queries-response.mapper.spec.ts
@@ -1,5 +1,5 @@
 import { PlatformNextQueriesResponse } from '../../../types/response.types';
-import { nextQueriesResponseMapper } from '../../response/next-queries-response.mapper';
+import { nextQueriesResponseMapper } from '../../responses/next-queries-response.mapper';
 
 describe('next queries response mapper', () => {
   it('should map the next queries', () => {
