+++ packages/x-adapter-platform/src/mappers/__tests__/response/search-response.mapper.spec.ts
@@ -1,4 +1,4 @@
-import { searchResponseMapper } from '../../response/search-response.mapper';
+import { searchResponseMapper } from '../../responses/search-response.mapper';
 import { PlatformSearchResponse } from '../../../types/response.types';
 import { platformSearchResponse } from '../../../__tests__/__fixtures__/platform-search.response';
 import { searchResponse } from '../../../__tests__/__fixtures__/search.response';
