+++ packages/x-adapter-platform/src/mappers/__tests__/response/sku-search-response.mapper.spec.ts
@@ -1,4 +1,4 @@
-import { skuSearchResponseMapper } from '../../response/sku-search-response.mapper';
+import { skuSearchResponseMapper } from '../../responses/sku-search-response.mapper';
 // eslint-disable-next-line max-len
 import { platformSkuSearchResponse } from '../../../__tests__/__fixtures__/platform-sku-search.response';
 import { skuSearchResponse } from '../../../__tests__/__fixtures__/sku-search.response';
