+++ packages/x-adapter-platform/src/mappers/index.ts
@@ -1,2 +1,2 @@
-export * from './request';
-export * from './response';
+export * from './requests';
+export * from './responses';
