+++ packages/x-adapter-platform/src/mappers/requests/query-suggestions-request.mapper.ts
@@ -1,9 +1,9 @@
 import { schemaMapperFactory } from '@empathyco/x-adapter-next';
 import { QuerySuggestionsRequest } from '@empathyco/x-types';
 // eslint-disable-next-line max-len
-import { querySuggestionsRequestMutableSchema } from '../../schemas/request/query-suggestions.schema';
+import { querySuggestionsRequestMutableSchema } from '../../schemas/requests/query-suggestions.schema';
 // eslint-disable-next-line max-len
-import { PlatformQuerySuggestionsRequest } from '../../types/request/query-suggestions-request.model';
+import { PlatformQuerySuggestionsRequest } from '../../types/requests/query-suggestions-request.model';
 
 export const querySuggestionsRequestMapper = schemaMapperFactory<
   QuerySuggestionsRequest,
