+++ packages/x-adapter-platform/src/mappers/responses/index.ts
