+++ packages/x-adapter-platform/src/mappers/responses/next-queries-response.mapper.ts
