+++ packages/x-adapter-platform/src/mappers/responses/query-suggestions-response.mapper.ts
@@ -1,9 +1,9 @@
 import { schemaMapperFactory } from '@empathyco/x-adapter-next';
 import { QuerySuggestionsResponse } from '@empathyco/x-types';
 // eslint-disable-next-line max-len
-import { querySuggestionsResponseMutableSchema } from '../../schemas/response/query-suggestions-response.schema';
+import { querySuggestionsResponseMutableSchema } from '../../schemas/responses/query-suggestions-response.schema';
 // eslint-disable-next-line max-len
-import { PlatformQuerySuggestionsResponse } from '../../types/response/query-suggestions-response.model';
+import { PlatformQuerySuggestionsResponse } from '../../types/responses/query-suggestions-response.model';
 
 export const querySuggestionsResponseMapper = schemaMapperFactory<
   PlatformQuerySuggestionsResponse,
