+++ packages/x-adapter-platform/src/mappers/responses/sku-search-response.mapper.ts
