+++ packages/x-adapter-platform/src/schemas/requests/query-suggestions.schema.ts
@@ -1,0 +1,13 @@
+import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
+import { QuerySuggestionsRequest } from '@empathyco/x-types';
+// eslint-disable-next-line max-len
+import { PlatformQuerySuggestionsRequest } from '../../types/requests/query-suggestions-request.model';
+
+export const querySuggestionsRequestMutableSchema = createMutableSchema<
+  Schema<QuerySuggestionsRequest, PlatformQuerySuggestionsRequest>
+>({
+  query: 'query',
+  start: 'start',
+  rows: 'rows',
+  extraParams: 'extraParams'
+});
