+++ packages/x-adapter-platform/src/schemas/responses/query-suggestions-response.schema.ts
@@ -1,0 +1,14 @@
+import { createMutableSchema, Schema } from '@empathyco/x-adapter-next';
+import { QuerySuggestionsResponse } from '@empathyco/x-types';
+// eslint-disable-next-line max-len
+import { PlatformQuerySuggestionsResponse } from '../../types/responses/query-suggestions-response.model';
+import { suggestionMutableSchema } from '../models/suggestion.schema';
+
+export const querySuggestionsResponseMutableSchema = createMutableSchema<
+  Schema<PlatformQuerySuggestionsResponse, QuerySuggestionsResponse>
+>({
+  suggestions: {
+    $path: 'topTrends.content',
+    $subSchema: suggestionMutableSchema
+  }
+});
