+++ packages/x-adapter-platform/src/types/models/suggestion.model.ts
