+++ packages/x-adapter-platform/src/types/requests/query-suggestions-request.model.ts
