+++ packages/x-adapter-platform/src/types/responses/index.ts
